feat(widgets): 테스트 실행 및 케이스 관리를 위한 주요 UI 컴포넌트 추가
- RunCaseDetailPanel을 포함한 실행 세부사항 관련 컴포넌트 추가
- SuiteCaseListSection 등 테스트 케이스 섹션 및 행 관리를 위한 컴포넌트 구현
- 상태 및 필터링 구성을 지원하는 공통 상수 및 유형 추가
