+++ src/view/project/runs/test-run-detail-view.tsx
@@ -7,14 +7,21 @@ import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
 import dynamic from 'next/dynamic';
 
 import { MainContainer } from '@/shared/lib/primitives';
-import { DSButton, RUN_STATUS_CONFIG, Skeleton, SkeletonCircle, EmptyState } from '@/shared/ui';
+import { RUN_STATUS_CONFIG } from '@/shared/ui';
 import { cn } from '@/shared/utils';
 import { useOutsideClick, useToggleSet } from '@/shared/hooks';
 import { type TestStatusData } from '@/widgets/project';
-import { testRunByIdQueryOptions, testRunsQueryOptions, updateTestCaseRunStatus, TestCaseRunDetail } from '@/features/runs';
+import { testRunByIdQueryOptions, testRunsQueryOptions, updateTestCaseRunStatus } from '@/features/runs';
 import { dashboardQueryOptions } from '@/features/dashboard';
 import { track, TESTRUN_EVENTS } from '@/shared/lib/analytics';
 import { ShareButton } from '@/features/runs-share/ui/share-button';
+import { ArrowLeft, XCircle, Keyboard } from 'lucide-react';
+
+import {
+  STATUS_CONFIG, SOURCE_TYPE_CONFIG,
+  RunDetailSkeleton, KeyboardShortcutsModal, RunCaseListPanel, RunCaseDetailPanel,
+  type StatusFilter, type TestCaseRunStatus, type GroupedCases,
+} from './_components';
 
 const TestStatusChart = dynamic(
   () => import('@/widgets/project/ui/test-status-chart').then(mod => ({ default: mod.TestStatusChart })),
@@ -24,78 +31,6 @@ const MilestoneGanttChart = dynamic(
   () => import('@/widgets/project/ui/milestone-gantt-chart').then(mod => ({ default: mod.MilestoneGanttChart })),
   { ssr: false, loading: () => <div className="bg-bg-2 rounded-[16px] p-6 h-[300px] animate-pulse" /> }
 );
-import {
-  ArrowLeft,
-  CheckCircle2,
-  XCircle,
-  AlertTriangle,
-  Circle,
-  Search,
-  ChevronRight,
-  ChevronDown,
-  MessageSquare,
-  Clock,
-  ListTodo,
-  PlayCircle,
-  FolderOpen,
-  Filter,
-  Keyboard,
-  Loader2,
-} from 'lucide-react';
-
-type StatusFilter = 'all' | 'untested' | 'pass' | 'fail' | 'blocked';
-type TestCaseRunStatus = 'untested' | 'pass' | 'fail' | 'blocked';
-
-interface GroupedCases {
-  groupKey: string;
-  suiteId: string | null;
-  suiteName: string;
-  cases: TestCaseRunDetail[];
-}
-
-const STATUS_CONFIG: Record<TestCaseRunStatus, {
-  label: string;
-  style: string;
-  bgStyle: string;
-  icon: React.ReactNode;
-  shortcut: string;
-}> = {
-  untested: {
-    label: 'Untested',
-    style: 'text-slate-400',
-    bgStyle: 'bg-slate-500/10 hover:bg-slate-500/20 border-slate-500/30',
-    icon: <Circle className="h-4 w-4" />,
-    shortcut: 'U',
-  },
-  pass: {
-    label: 'Pass',
-    style: 'text-green-400',
-    bgStyle: 'bg-green-500/10 hover:bg-green-500/20 border-green-500/30',
-    icon: <CheckCircle2 className="h-4 w-4" />,
-    shortcut: 'P',
-  },
-  fail: {
-    label: 'Fail',
-    style: 'text-red-400',
-    bgStyle: 'bg-red-500/10 hover:bg-red-500/20 border-red-500/30',
-    icon: <XCircle className="h-4 w-4" />,
-    shortcut: 'F',
-  },
-  blocked: {
-    label: 'Blocked',
-    style: 'text-amber-400',
-    bgStyle: 'bg-amber-500/10 hover:bg-amber-500/20 border-amber-500/30',
-    icon: <AlertTriangle className="h-4 w-4" />,
-    shortcut: 'B',
-  },
-};
-
-
-const SOURCE_TYPE_CONFIG: Record<string, { label: string; icon: React.ReactNode }> = {
-  SUITE: { label: '스위트', icon: <ListTodo className="h-4 w-4" /> },
-  MILESTONE: { label: '마일스톤', icon: <Clock className="h-4 w-4" /> },
-  ADHOC: { label: '직접 선택', icon: <PlayCircle className="h-4 w-4" /> },
-};
 
 export const TestRunDetailView = () => {
   const params = useParams();
@@ -164,7 +99,6 @@ export const TestRunDetailView = () => {
       untested: testRun.stats.untested,
     };
   }, [testRun]);
-
   // 테스트 실행 상세 View 이벤트
   useEffect(() => {
     if (testRun) {
@@ -324,53 +258,7 @@ export const TestRunDetailView = () => {
 
   // 로딩 상태 — 스켈레톤 UI
   if (isLoading) {
-    return (
-      <MainContainer className="flex flex-1 flex-col min-h-screen">
-        {/* Header skeleton */}
-        <div className="border-line-2 flex items-center justify-between border-b px-6 py-4">
-          <div className="flex items-center gap-4">
-            <Skeleton className="h-4 w-4" />
-            <div className="flex flex-col gap-1">
-              <Skeleton className="h-6 w-48" />
-              <Skeleton className="h-4 w-32" />
-            </div>
-          </div>
-          <div className="flex items-center gap-4">
-            <div className="flex items-center gap-3">
-              {Array.from({ length: 4 }).map((_, i) => (
-                <Skeleton key={i} className="h-5 w-12" />
-              ))}
-            </div>
-            <SkeletonCircle className="h-2 w-32" />
-          </div>
-        </div>
-        {/* Content skeleton */}
-        <div className="sticky top-0 flex h-screen overflow-hidden">
-          <div className="border-line-2 flex w-[60%] flex-col border-r">
-            <div className="border-line-2 flex flex-col gap-3 border-b p-4">
-              <Skeleton className="h-10 w-full rounded-lg border border-line-2 bg-bg-2" />
-              <div className="flex gap-2">
-                <Skeleton className="h-10 flex-1 rounded-lg border border-line-2 bg-bg-2" />
-                <Skeleton className="h-10 flex-1 rounded-lg border border-line-2 bg-bg-2" />
-              </div>
-            </div>
-            {Array.from({ length: 8 }).map((_, i) => (
-              <div key={i} className="border-line-2 flex items-center gap-3 border-b py-3 pl-8 pr-4">
-                <SkeletonCircle className="h-4 w-4 shrink-0" />
-                <div className="flex-1 flex flex-col gap-1">
-                  <Skeleton className="h-3 w-16" />
-                  <Skeleton className="h-4 w-48" />
-                </div>
-              </div>
-            ))}
-          </div>
-          <div className="flex w-[40%] flex-col items-center justify-center">
-            <Skeleton className="h-12 w-12 mb-4" />
-            <Skeleton className="h-5 w-48" />
-          </div>
-        </div>
-      </MainContainer>
-    );
+    return <RunDetailSkeleton />;
   }
 
   if (isError || !testRun) {
@@ -464,470 +352,47 @@ export const TestRunDetailView = () => {
 
         {/* Main Content - sticky panels */}
         <div className="sticky top-0 flex h-screen overflow-hidden">
-          {/* Left Panel - Test Case List */}
-          <div className="border-line-2 flex w-[60%] flex-col border-r">
-            {/* Search & Filter */}
-            <div className="border-line-2 flex flex-col gap-3 border-b p-4">
-              <div className="relative">
-                <Search className="text-text-3 pointer-events-none absolute left-3 top-1/2 h-4 w-4 -translate-y-1/2" />
-                <input
-                  type="text"
-                  placeholder="케이스 검색..."
-                  value={searchQuery}
-                  onChange={(e) => setSearchQuery(e.target.value)}
-                  className="bg-bg-2 border-line-2 text-text-1 placeholder:text-text-4 focus:border-primary w-full rounded-lg border py-2 pl-10 pr-4 text-sm focus:outline-none focus:ring-1 focus:ring-primary"
-                />
-              </div>
+          <RunCaseListPanel
+            searchQuery={searchQuery}
+            setSearchQuery={setSearchQuery}
+            statusFilter={statusFilter}
+            setStatusFilter={setStatusFilter}
+            suiteFilter={suiteFilter}
+            setSuiteFilter={setSuiteFilter}
+            availableSuites={availableSuites}
+            filteredCases={filteredCases}
+            groupedCases={groupedCases}
+            collapsedGroups={collapsedGroups}
+            selectedCaseId={selectedCaseId}
+            setSelectedCaseId={setSelectedCaseId}
+            testRunStats={testRun.stats}
+            showStatusFilterDropdown={showStatusFilterDropdown}
+            setShowStatusFilterDropdown={setShowStatusFilterDropdown}
+            showSuiteFilterDropdown={showSuiteFilterDropdown}
+            setShowSuiteFilterDropdown={setShowSuiteFilterDropdown}
+            statusFilterRef={statusFilterRef}
+            suiteFilterRef={suiteFilterRef}
+          />
 
-              {/* Filter Dropdowns */}
-              <div className="flex gap-2">
-                {/* Status Filter Dropdown */}
-                <div className="relative flex-1" ref={statusFilterRef}>
-                  <button
-                    onClick={() => {
-                      setShowStatusFilterDropdown(!showStatusFilterDropdown);
-                      setShowSuiteFilterDropdown(false);
-                    }}
-                    className="bg-bg-2 border-line-2 hover:border-line-1 flex w-full items-center justify-between rounded-lg border px-3 py-2 text-sm transition-colors"
-                  >
-                    <div className="flex items-center gap-2">
-                      {statusFilter === 'all' ? (
-                        <>
-                          <Filter className="text-text-3 h-4 w-4" />
-                          <span className="text-text-1">전체 상태</span>
-                        </>
-                      ) : (
-                        <>
-                          <span className={STATUS_CONFIG[statusFilter].style}>
-                            {STATUS_CONFIG[statusFilter].icon}
-                          </span>
-                          <span className="text-text-1">{STATUS_CONFIG[statusFilter].label}</span>
-                          <span className="text-text-3 text-xs">({testRun.stats[statusFilter]})</span>
-                        </>
-                      )}
-                    </div>
-                    <ChevronDown className={cn('text-text-3 h-4 w-4 transition-transform', showStatusFilterDropdown && 'rotate-180')} />
-                  </button>
-
-                  {showStatusFilterDropdown && (
-                    <div className="bg-bg-2 border-line-2 absolute left-0 right-0 top-full z-30 mt-1 overflow-hidden rounded-lg border shadow-xl">
-                      <button
-                        onClick={() => {
-                          setStatusFilter('all');
-                          setShowStatusFilterDropdown(false);
-                        }}
-                        className={cn(
-                          'flex w-full items-center gap-2 px-3 py-2 text-sm transition-colors',
-                          statusFilter === 'all' ? 'bg-primary/10 text-primary' : 'hover:bg-bg-3'
-                        )}
-                      >
-                        <Filter className="h-4 w-4" />
-                        <span>전체 상태</span>
-                        <span className="text-text-3 ml-auto text-xs">({testRun.stats.total})</span>
-                      </button>
-                      {(['untested', 'pass', 'fail', 'blocked'] as const).map((status) => (
-                        <button
-                          key={status}
-                          onClick={() => {
-                            setStatusFilter(status);
-                            setShowStatusFilterDropdown(false);
-                          }}
-                          className={cn(
-                            'flex w-full items-center gap-2 px-3 py-2 text-sm transition-colors',
-                            statusFilter === status ? 'bg-primary/10 text-primary' : 'hover:bg-bg-3'
-                          )}
-                        >
-                          <span className={STATUS_CONFIG[status].style}>{STATUS_CONFIG[status].icon}</span>
-                          <span>{STATUS_CONFIG[status].label}</span>
-                          <span className="text-text-3 ml-auto text-xs">({testRun.stats[status]})</span>
-                        </button>
-                      ))}
-                    </div>
-                  )}
-                </div>
-
-                {/* Suite Filter Dropdown */}
-                <div className="relative flex-1" ref={suiteFilterRef}>
-                  <button
-                    onClick={() => {
-                      setShowSuiteFilterDropdown(!showSuiteFilterDropdown);
-                      setShowStatusFilterDropdown(false);
-                    }}
-                    className="bg-bg-2 border-line-2 hover:border-line-1 flex w-full items-center justify-between rounded-lg border px-3 py-2 text-sm transition-colors"
-                  >
-                    <div className="flex items-center gap-2">
-                      <FolderOpen className={cn('h-4 w-4', suiteFilter !== 'all' ? 'text-primary' : 'text-text-3')} />
-                      <span className="text-text-1 truncate">
-                        {suiteFilter === 'all'
-                          ? '스위트'
-                          : availableSuites.find(s => s.id === suiteFilter)?.name || '스위트'}
-                      </span>
-                    </div>
-                    <ChevronDown className={cn('text-text-3 h-4 w-4 flex-shrink-0 transition-transform', showSuiteFilterDropdown && 'rotate-180')} />
-                  </button>
-
-                  {showSuiteFilterDropdown && (
-                    <div className="bg-bg-2 border-line-2 absolute left-0 right-0 top-full z-30 mt-1 max-h-60 overflow-y-auto rounded-lg border shadow-xl">
-                      <button
-                        onClick={() => {
-                          setSuiteFilter('all');
-                          setShowSuiteFilterDropdown(false);
-                        }}
-                        className={cn(
-                          'flex w-full items-center gap-2 px-3 py-2 text-sm transition-colors',
-                          suiteFilter === 'all' ? 'bg-primary/10 text-primary' : 'hover:bg-bg-3'
-                        )}
-                      >
-                        <FolderOpen className="h-4 w-4" />
-                        <span>전체 스위트</span>
-                      </button>
-                      {availableSuites.length === 0 ? (
-                        <div className="text-text-3 px-3 py-2 text-sm">스위트 없음</div>
-                      ) : (
-                        availableSuites.map((suite) => (
-                          <button
-                            key={suite.id}
-                            onClick={() => {
-                              setSuiteFilter(suite.id);
-                              setShowSuiteFilterDropdown(false);
-                            }}
-                            className={cn(
-                              'flex w-full items-center gap-2 px-3 py-2 text-sm transition-colors',
-                              suiteFilter === suite.id ? 'bg-primary/10 text-primary' : 'hover:bg-bg-3'
-                            )}
-                          >
-                            <FolderOpen className="h-4 w-4" />
-                            <span className="truncate">{suite.name}</span>
-                          </button>
-                        ))
-                      )}
-                    </div>
-                  )}
-                </div>
-              </div>
-            </div>
-
-            {/* Case List - Grouped by Source */}
-            <div className="flex-1 overflow-y-auto">
-              {filteredCases.length === 0 ? (
-                <div className="flex flex-col items-center justify-center py-12 text-center">
-                  <Filter className="text-text-3 mb-2 h-8 w-8" />
-                  <p className="text-text-2">검색 결과가 없습니다.</p>
-                </div>
-              ) : (
-                groupedCases.map((group) => {
-                  const isCollapsed = collapsedGroups.has(group.groupKey);
-                  const groupPass = group.cases.filter(c => c.status === 'pass').length;
-                  const groupFail = group.cases.filter(c => c.status === 'fail').length;
-                  const groupBlocked = group.cases.filter(c => c.status === 'blocked').length;
-
-                  return (
-                    <div key={group.groupKey}>
-                      {/* Suite Group Header - non-interactive section divider */}
-                      <div
-                        role="button"
-                        tabIndex={0}
-                        onClick={() => collapsedGroups.toggle(group.groupKey)}
-                        onKeyDown={(e) => { if (e.key === 'Enter' || e.key === ' ') { e.preventDefault(); collapsedGroups.toggle(group.groupKey); } }}
-                        className="bg-bg-3/50 border-line-2 sticky top-0 z-10 flex cursor-pointer select-none items-center gap-2 border-b px-3 py-1.5"
-                      >
-                        <ChevronDown
-                          className={cn(
-                            'text-text-4 h-3.5 w-3.5 transition-transform',
-                            isCollapsed && '-rotate-90'
-                          )}
-                        />
-                        <FolderOpen className="text-text-3 h-3.5 w-3.5" />
-                        <span className="text-text-2 flex-1 text-xs font-semibold uppercase tracking-wide">
-                          {group.suiteName}
-                        </span>
-                        <div className="flex items-center gap-1.5 text-[10px]">
-                          {groupPass > 0 && (
-                            <span className="text-green-400/70">{groupPass}P</span>
-                          )}
-                          {groupFail > 0 && (
-                            <span className="text-red-400/70">{groupFail}F</span>
-                          )}
-                          {groupBlocked > 0 && (
-                            <span className="text-amber-400/70">{groupBlocked}B</span>
-                          )}
-                          <span className="text-text-4">
-                            {group.cases.length}
-                          </span>
-                        </div>
-                      </div>
-
-                      {/* Test Case Items */}
-                      {!isCollapsed && group.cases.map((tc, index) => {
-                        const config = STATUS_CONFIG[tc.status];
-                        const isSelected = tc.id === selectedCaseId;
-
-                        return (
-                          <button
-                            key={tc.id}
-                            onClick={() => setSelectedCaseId(tc.id)}
-                            className={cn(
-                              'border-line-2 flex w-full items-center gap-3 border-b py-3 pl-8 pr-4 text-left transition-colors',
-                              isSelected ? 'bg-primary/10' : 'hover:bg-bg-2'
-                            )}
-                          >
-                            <span className={cn('flex-shrink-0', config.style)}>{config.icon}</span>
-                            <div className="min-w-0 flex-1">
-                              <div className="flex items-center gap-2">
-                                <span className="text-primary font-mono text-xs">{tc.code || `#${index + 1}`}</span>
-                                {tc.comment && <MessageSquare className="text-text-3 h-3 w-3" />}
-                              </div>
-                              <p className="text-text-1 truncate text-sm">{tc.title || '제목 없음'}</p>
-                            </div>
-                            {isSelected && <ChevronRight className="text-text-3 h-4 w-4 flex-shrink-0" />}
-                          </button>
-                        );
-                      })}
-                    </div>
-                  );
-                })
-              )}
-            </div>
-          </div>
-
-          {/* Right Panel - Test Case Detail */}
-          <div className="flex w-[40%] flex-col overflow-y-auto">
-            {selectedCase ? (
-              <>
-                {/* Case Header */}
-                <div className="border-line-2 border-b p-6">
-                  <div className="mb-2 flex items-center gap-2">
-                    <span className="text-primary font-mono text-sm">{selectedCase.code}</span>
-                    <span className={cn(
-                      'inline-flex items-center gap-1 rounded-full px-2 py-0.5 text-xs font-medium',
-                      STATUS_CONFIG[selectedCase.status].style,
-                      STATUS_CONFIG[selectedCase.status].bgStyle.split(' ')[0]
-                    )}>
-                      {STATUS_CONFIG[selectedCase.status].icon}
-                      {STATUS_CONFIG[selectedCase.status].label}
-                    </span>
-                  </div>
-                  <h2 className="text-text-1 text-xl font-semibold">{selectedCase.title || '제목 없음'}</h2>
-                </div>
-
-                {/* Status Dropdown */}
-                <div className="border-line-2 relative border-b p-6">
-                  {updateMutation.isPending && (
-                    <div className="absolute inset-0 z-10 flex items-center justify-center bg-bg-1/70 backdrop-blur-sm">
-                      <div className="flex items-center gap-2 text-primary">
-                        <Loader2 className="h-5 w-5 animate-spin" />
-                        <span className="text-sm font-medium">저장 중...</span>
-                      </div>
-                    </div>
-                  )}
-                  <h3 className="text-text-2 mb-3 text-sm font-medium">결과 기록</h3>
-                  <div className="relative" ref={statusDropdownRef}>
-                    <button
-                      onClick={() => setShowStatusDropdown(!showStatusDropdown)}
-                      disabled={updateMutation.isPending}
-                      className={cn(
-                        'flex w-full items-center justify-between rounded-lg border-2 px-4 py-3 transition-all',
-                        STATUS_CONFIG[selectedCase.status].bgStyle,
-                        STATUS_CONFIG[selectedCase.status].style
-                      )}
-                    >
-                      <div className="flex items-center gap-3">
-                        <span className="text-lg">{STATUS_CONFIG[selectedCase.status].icon}</span>
-                        <span className="font-medium">{STATUS_CONFIG[selectedCase.status].label}</span>
-                      </div>
-                      <ChevronDown className={cn('h-5 w-5 transition-transform', showStatusDropdown && 'rotate-180')} />
-                    </button>
-
-                    {showStatusDropdown && (
-                      <div className="bg-bg-2 border-line-2 absolute left-0 right-0 top-full z-20 mt-1 overflow-hidden rounded-lg border shadow-xl">
-                        {(['pass', 'fail', 'blocked', 'untested'] as const).map((status) => {
-                          const config = STATUS_CONFIG[status];
-                          const isActive = selectedCase.status === status;
-
-                          return (
-                            <button
-                              key={status}
-                              onClick={() => {
-                                handleStatusChange(status);
-                                setShowStatusDropdown(false);
-                              }}
-                              className={cn(
-                                'flex w-full items-center gap-3 px-4 py-3 transition-colors',
-                                isActive ? cn(config.bgStyle, config.style) : 'hover:bg-bg-3'
-                              )}
-                            >
-                              <span className={cn('text-lg', isActive ? config.style : 'text-text-3')}>
-                                {config.icon}
-                              </span>
-                              <span className={cn('flex-1 text-left font-medium', isActive ? config.style : 'text-text-1')}>
-                                {config.label}
-                              </span>
-                              <kbd className="bg-bg-3 text-text-4 rounded px-1.5 py-0.5 text-xs">
-                                {config.shortcut}
-                              </kbd>
-                            </button>
-                          );
-                        })}
-                      </div>
-                    )}
-                  </div>
-                </div>
-
-                {/* Test Case Details */}
-                <div className="border-line-2 flex flex-col gap-4 border-b p-6">
-                  {selectedCase.preCondition && (
-                    <div>
-                      <h3 className="text-text-2 mb-2 text-sm font-medium">사전조건</h3>
-                      <div className="bg-bg-2 rounded-lg p-3">
-                        <p className="text-text-1 whitespace-pre-wrap text-sm">{selectedCase.preCondition}</p>
-                      </div>
-                    </div>
-                  )}
-                  {selectedCase.steps && (
-                    <div>
-                      <h3 className="text-text-2 mb-2 text-sm font-medium">테스트 스텝</h3>
-                      <div className="bg-bg-2 rounded-lg p-3">
-                        <p className="text-text-1 whitespace-pre-wrap text-sm">{selectedCase.steps}</p>
-                      </div>
-                    </div>
-                  )}
-                  {selectedCase.expectedResult && (
-                    <div>
-                      <h3 className="text-text-2 mb-2 text-sm font-medium">기대 결과</h3>
-                      <div className="bg-bg-2 rounded-lg p-3">
-                        <p className="text-text-1 whitespace-pre-wrap text-sm">{selectedCase.expectedResult}</p>
-                      </div>
-                    </div>
-                  )}
-                  {!selectedCase.preCondition && !selectedCase.steps && !selectedCase.expectedResult && (
-                    <p className="text-text-3 text-sm">사전조건, 테스트 스텝, 기대 결과가 등록되지 않았습니다.</p>
-                  )}
-                </div>
-
-                {/* Comment Section */}
-                <div className="flex-1 p-6">
-                  <h3 className="text-text-2 mb-3 text-sm font-medium">코멘트</h3>
-                  <textarea
-                    value={comment}
-                    onChange={(e) => setComment(e.target.value)}
-                    placeholder="테스트 결과에 대한 코멘트를 입력하세요..."
-                    className="bg-bg-2 border-line-2 text-text-1 placeholder:text-text-4 focus:border-primary h-32 w-full resize-none rounded-lg border p-3 text-sm focus:outline-none focus:ring-1 focus:ring-primary"
-                  />
-                  <div className="mt-3 flex justify-end">
-                    <DSButton
-                      size="small"
-                      variant="ghost"
-                      onClick={() => {
-                        if (selectedCase) {
-                          updateMutation.mutate({
-                            testCaseRunId: selectedCase.id,
-                            status: selectedCase.status,
-                            comment: comment || null,
-                          });
-                        }
-                      }}
-                      disabled={updateMutation.isPending || comment === (selectedCase?.comment || '')}
-                      className="flex items-center gap-2"
-                    >
-                      {updateMutation.isPending && <Loader2 className="h-3 w-3 animate-spin" />}
-                      코멘트 저장
-                    </DSButton>
-                  </div>
-                </div>
-
-                {/* Navigation Footer */}
-                <div className="border-line-2 flex items-center justify-between border-t p-4">
-                  <div className="text-text-3 text-sm">
-                    {filteredCases.findIndex((tc) => tc.id === selectedCaseId) + 1} / {filteredCases.length}
-                  </div>
-                  <div className="flex gap-2">
-                    <DSButton
-                      size="small"
-                      variant="ghost"
-                      onClick={() => {
-                        const idx = filteredCases.findIndex((tc) => tc.id === selectedCaseId);
-                        if (idx > 0) setSelectedCaseId(filteredCases[idx - 1].id);
-                      }}
-                      disabled={filteredCases.findIndex((tc) => tc.id === selectedCaseId) === 0}
-                    >
-                      이전
-                    </DSButton>
-                    <DSButton
-                      size="small"
-                      variant="ghost"
-                      onClick={() => {
-                        const idx = filteredCases.findIndex((tc) => tc.id === selectedCaseId);
-                        if (idx < filteredCases.length - 1) setSelectedCaseId(filteredCases[idx + 1].id);
-                      }}
-                      disabled={filteredCases.findIndex((tc) => tc.id === selectedCaseId) === filteredCases.length - 1}
-                    >
-                      다음
-                    </DSButton>
-                  </div>
-                </div>
-              </>
-            ) : (
-              <EmptyState
-                icon={<ListTodo className="h-12 w-12" />}
-                title="테스트 케이스를 선택하세요"
-                description="왼쪽 목록에서 테스트할 케이스를 선택합니다."
-                className="flex-1"
-              />
-            )}
-          </div>
+          <RunCaseDetailPanel
+            selectedCase={selectedCase}
+            filteredCases={filteredCases}
+            selectedCaseId={selectedCaseId}
+            setSelectedCaseId={setSelectedCaseId}
+            comment={comment}
+            setComment={setComment}
+            handleStatusChange={handleStatusChange}
+            updateMutation={updateMutation}
+            showStatusDropdown={showStatusDropdown}
+            setShowStatusDropdown={setShowStatusDropdown}
+            statusDropdownRef={statusDropdownRef}
+          />
         </div>
 
-      {/* Keyboard Shortcuts Modal */}
-      {showShortcuts && (
-        <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/50" onClick={() => setShowShortcuts(false)}>
-          <div className="bg-bg-2 border-line-2 w-96 rounded-2xl border p-6 shadow-xl" onClick={(e) => e.stopPropagation()}>
-            <h3 className="text-text-1 mb-4 text-lg font-semibold">키보드 단축키</h3>
-            <div className="space-y-3">
-              <div className="flex items-center justify-between">
-                <span className="text-text-2">Pass로 기록</span>
-                <kbd className="bg-bg-3 text-text-1 rounded px-2 py-1 text-sm">P</kbd>
-              </div>
-              <div className="flex items-center justify-between">
-                <span className="text-text-2">Fail로 기록</span>
-                <kbd className="bg-bg-3 text-text-1 rounded px-2 py-1 text-sm">F</kbd>
-              </div>
-              <div className="flex items-center justify-between">
-                <span className="text-text-2">Blocked로 기록</span>
-                <kbd className="bg-bg-3 text-text-1 rounded px-2 py-1 text-sm">B</kbd>
-              </div>
-              <div className="flex items-center justify-between">
-                <span className="text-text-2">Untested로 초기화</span>
-                <kbd className="bg-bg-3 text-text-1 rounded px-2 py-1 text-sm">U</kbd>
-              </div>
-              <div className="border-line-2 my-3 border-t" />
-              <div className="flex items-center justify-between">
-                <span className="text-text-2">다음 케이스</span>
-                <div className="flex gap-1">
-                  <kbd className="bg-bg-3 text-text-1 rounded px-2 py-1 text-sm">↓</kbd>
-                  <kbd className="bg-bg-3 text-text-1 rounded px-2 py-1 text-sm">J</kbd>
-                </div>
-              </div>
-              <div className="flex items-center justify-between">
-                <span className="text-text-2">이전 케이스</span>
-                <div className="flex gap-1">
-                  <kbd className="bg-bg-3 text-text-1 rounded px-2 py-1 text-sm">↑</kbd>
-                  <kbd className="bg-bg-3 text-text-1 rounded px-2 py-1 text-sm">K</kbd>
-                </div>
-              </div>
-              <div className="flex items-center justify-between">
-                <span className="text-text-2">단축키 도움말</span>
-                <kbd className="bg-bg-3 text-text-1 rounded px-2 py-1 text-sm">?</kbd>
-              </div>
-            </div>
-            <button
-              onClick={() => setShowShortcuts(false)}
-              className="bg-primary mt-6 w-full rounded-lg py-2 text-sm font-medium text-white transition-colors hover:bg-primary/90"
-            >
-              닫기
-            </button>
-          </div>
-        </div>
-      )}
+      <KeyboardShortcutsModal
+        open={showShortcuts}
+        onClose={() => setShowShortcuts(false)}
+      />
 
     </MainContainer>
   );
